refactor: Merge ObjectManger functionality into Vocabulary object
