--- game/Assets/ObjectManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e974dac0f69f90b4fa12c047fb7c4236
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
